pyproject.toml
[project]
name = "kilko-waste"
version = "0.1.0"
description = "Kilko Container Management API Client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "python-dateutil>=2.9.0.post0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
